passes compilation.. now onto linking
Former-commit-id: acb091cd49414991a623b751dee31be827b85c76 [formerly 40d11cf006c3b8b7fab660eab63871e7f6a5a4ae]
Former-commit-id: 0b34f177627f366548d8d49e26421ad383a6541f

+++ src/stack.h
@@ -94,7 +94,7 @@ namespace FT
         float * dev_f; 
         bool * dev_b; 
         std::map<char, size_t> idx; 
-        int N; 
+        size_t N; 
 
         Stacks()
         {
